[profiler] Split up profiler test file
